feat:add overlay suggested seed list FA-58
