+++ lib/auth/onboard_import_account_page/onboard_import_account_page.dart
@@ -1,9 +1,8 @@
 // ignore_for_file: lines_longer_than_80_chars, prefer_final_locals
 
-import 'dart:convert';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:ui_library/ui_library_export.dart';
+import 'package:uplink/auth/onboard_import_account_page/models/seed_text_field.dart';
 import 'package:uplink/linking_satelittes_page.dart';
 
 class OnboardImportAccountPage extends StatefulWidget {
@@ -15,41 +14,21 @@ class OnboardImportAccountPage extends StatefulWidget {
 }
 
 class _OnboardImportAccountPageState extends State<OnboardImportAccountPage> {
-  final controller = TextEditingController();
-  final focusNode = FocusNode();
-  late List<String> _bip39Dic;
   List<String> selectedPassphraseList = [];
-  List<String> suggestedPassphraseList = [];
-  int? tapedWordIndex; //the index for taped word in suggestedPassphraseList
+  // List<String> suggestedPassphraseList = [];
+  // int? tapedWordIndex; //the index for taped word in suggestedPassphraseList
   String query = ''; //query value from the text field
-  bool showSuggestedWordsGridView = false;
+
   bool selectionFinished = false;
   bool isWrongSeeds = false;
 
   @override
   void initState() {
     super.initState();
-    _loadBip39().then((value) {
-      _bip39Dic = value;
-      suggestedPassphraseList.addAll(_bip39Dic);
-    });
-    focusNode.addListener(() {
-      if (focusNode.hasFocus) {
-        setState(() {
-          showSuggestedWordsGridView = true;
-        });
-      } else {
-        setState(() {
-          showSuggestedWordsGridView = false;
-        });
-      }
-    });
   }
 
   @override
   void dispose() {
-    controller.dispose();
-    focusNode.dispose();
     super.dispose();
   }
 
@@ -72,89 +51,93 @@ class _OnboardImportAccountPageState extends State<OnboardImportAccountPage> {
               ),
               const SizedBox(height: 16),
 //TextField
-              Container(
-                height: 48,
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(4),
-                  color: UColors.foregroundDark,
-                ),
-                padding: const EdgeInsets.symmetric(horizontal: 16),
-                child: TextField(
-                  autofocus: true,
-                  focusNode: focusNode,
-                  autocorrect: false,
-                  controller: controller,
-                  cursorColor: Colors.white,
-                  style: UTextStyle.H5_fifthHeader.style
-                      .returnTextStyleType()
-                      .copyWith(color: Colors.white),
-                  decoration: InputDecoration(
-                    border: InputBorder.none,
-                    hintText: 'Enter Passphrase',
-                    hintStyle: UTextStyle.H5_fifthHeader.style
-                        .returnTextStyleType()
-                        .copyWith(color: UColors.textDark),
-                  ),
-                  textInputAction: TextInputAction.newline,
-                  onChanged: (word) {
-                    //update suggestedPassphraseList to update the suggestion memu
-                    searchPassphrass(controller.text.toLowerCase());
-                  },
-                  onSubmitted: (passphrase) {
-                    addInSelectedWordGridView(
-                      passphrase: passphrase,
-                      fromTextField: true,
-                    );
-                  },
-                ),
+              SeedTextField(
+                selectedPassphraseList: selectedPassphraseList,
+                selectionFinished: selectionFinished,
               ),
+              // Container(
+              //   height: 48,
+              //   decoration: BoxDecoration(
+              //     borderRadius: BorderRadius.circular(4),
+              //     color: UColors.foregroundDark,
+              //   ),
+              //   padding: const EdgeInsets.symmetric(horizontal: 16),
+              //   child: TextField(
+              //     autofocus: true,
+              //     focusNode: focusNode,
+              //     autocorrect: false,
+              //     controller: controller,
+              //     cursorColor: Colors.white,
+              //     style: UTextStyle.H5_fifthHeader.style
+              //         .returnTextStyleType()
+              //         .copyWith(color: Colors.white),
+              //     decoration: InputDecoration(
+              //       border: InputBorder.none,
+              //       hintText: 'Enter Passphrase',
+              //       hintStyle: UTextStyle.H5_fifthHeader.style
+              //           .returnTextStyleType()
+              //           .copyWith(color: UColors.textDark),
+              //     ),
+              //     textInputAction: TextInputAction.newline,
+              //     onChanged: (word) {
+              //       //update suggestedPassphraseList to update the suggestion memu
+              //       searchPassphrass(controller.text.toLowerCase());
+              //     },
+              //     onSubmitted: (passphrase) {
+              //       addInSelectedWordGridView(
+              //         passphrase: passphrase,
+              //         fromTextField: true,
+              //       );
+              //     },
+              //   ),
+              // ),
               const SizedBox(height: 8),
 //suggested list
-              if (suggestedPassphraseList.isNotEmpty &&
-                  showSuggestedWordsGridView)
-                ConstrainedBox(
-                  constraints: const BoxConstraints(
-                    maxHeight: 230, //240
-                  ),
-                  child: ListView.builder(
-                    //Changing the size by its content
-                    shrinkWrap: true,
-                    physics: const ClampingScrollPhysics(),
-                    itemCount: suggestedPassphraseList.length,
-                    itemBuilder: (context, index) {
-                      final passphrase = suggestedPassphraseList[index];
-                      var highlight = false;
-                      if (tapedWordIndex == index) highlight = true;
-                      return InkWell(
-                        child: Container(
-                          color: highlight
-                              ? UColors.ctaBlue
-                              : UColors.foregroundDark,
-                          height: 48,
-                          padding: const EdgeInsets.symmetric(
-                            horizontal: 16,
-                            vertical: 16,
-                          ),
-                          child: UText(
-                            passphrase,
-                            textStyle: UTextStyle.BUT1_primaryButton,
-                          ),
-                        ),
-                        onTap: () {
-                          setState(() {
-                            tapedWordIndex = index;
-                            addInSelectedWordGridView(
-                              passphrase: passphrase,
-                              fromTextField: false,
-                            );
-                          });
-                        },
-                      );
-                    },
-                  ),
-                )
-              else
-                const SizedBox(),
+              // if (suggestedPassphraseList.isNotEmpty &&
+              //     showSuggestedWordsGridView)
+              //   ConstrainedBox(
+              //     constraints: const BoxConstraints(
+              //       maxHeight: 230, //240
+              //     ),
+              //     child: ListView.builder(
+              //       //Changing the size by its content
+              //       shrinkWrap: true,
+              //       physics: const ClampingScrollPhysics(),
+              //       itemCount: suggestedPassphraseList.length,
+              //       itemBuilder: (context, index) {
+              //         final passphrase = suggestedPassphraseList[index];
+              //         var highlight = false;
+              //         if (tapedWordIndex == index) highlight = true;
+              //         return InkWell(
+              //           child: Container(
+              //             color: highlight
+              //                 ? UColors.ctaBlue
+              //                 : UColors.foregroundDark,
+              //             height: 48,
+              //             padding: const EdgeInsets.symmetric(
+              //               horizontal: 16,
+              //               vertical: 16,
+              //             ),
+              //             child: UText(
+              //               passphrase,
+              //               textStyle: UTextStyle.BUT1_primaryButton,
+              //             ),
+              //           ),
+              //           onTap: () {
+              //             setState(() {
+              //               tapedWordIndex = index;
+              //               addInSelectedWordGridView(
+              //                 passphrase: passphrase,
+              //                 fromTextField: false,
+              //               );
+              //             });
+              //           },
+              //         );
+              //       },
+              //     ),
+              //   )
+              // else
+              //   const SizedBox(),
               const SizedBox(height: 24),
 //Selected words grid view
               CustomScrollView(
@@ -196,13 +179,6 @@ class _OnboardImportAccountPageState extends State<OnboardImportAccountPage> {
                 ),
 //Button to recovery account
               const SizedBox(height: 24),
-              // if (isWrongSeeds == true)
-              //   UButton.primary(
-              //     label: 'Try Again',
-              //     uIconData: UIcons.refresh_try_again,
-              //     onPressed: () {},
-              //   )
-              // else
               Opacity(
                 opacity: selectionFinished ? 1 : 0.5,
                 child: UButton.primary(
@@ -237,70 +213,5 @@ class _OnboardImportAccountPageState extends State<OnboardImportAccountPage> {
         ),
       ),
     );
-  }
-
-  Future<List<String>> _loadBip39() async {
-    var wordsList = <String>[];
-    await rootBundle.loadString('assets/bip39_english.txt').then((word) {
-      for (final value in const LineSplitter().convert(word)) {
-        wordsList.add(value);
-      }
-    });
-    return wordsList;
-  }
-
-  void searchPassphrass(String query) {
-    final _tempDicList = <String>[..._bip39Dic];
-
-    if (query.isNotEmpty) {
-      final _tempSuggestList = <String>[];
-      //add any word start with [query] into _tempSuggestList
-      for (final passPhrase in _tempDicList) {
-        if (passPhrase.toLowerCase().startsWith(query)) {
-          _tempSuggestList.add(passPhrase);
-        }
-      }
-
-      //clear previous suggestedPassphraseList
-      //update it with new suggestedPassphraseList
-      setState(() {
-        suggestedPassphraseList
-          ..clear()
-          ..addAll(_tempSuggestList);
-      });
-    } else {
-      //when the text from the textfield is null
-      //clear previous suggestedPassphraseList list
-      //add all bip39 words
-      setState(() {
-        suggestedPassphraseList
-          ..clear()
-          ..addAll(_tempDicList);
-      });
-    }
-  }
-
-  void addInSelectedWordGridView({
-    required String passphrase,
-    required bool fromTextField,
-  }) {
-    if (_bip39Dic.contains(passphrase)) {
-      if (selectedPassphraseList.length < 12) {
-        //list hasn't full
-        setState(() {
-          selectedPassphraseList.add(passphrase);
-          if (selectedPassphraseList.length == 12) selectionFinished = true;
-          if (fromTextField == true) {
-            controller.clear();
-            suggestedPassphraseList
-              ..clear()
-              ..addAll(_bip39Dic);
-            focusNode.requestFocus();
-          }
-        });
-      } else {
-        // TODO(yijing): list is full
-      }
-    }
   }
 }
